Initial 1.21.5 commit.

+++ src/main/java/vectorwing/farmersdelight/common/block/RichSoilFarmlandBlock.java
@@ -114,7 +114,7 @@ public class RichSoilFarmlandBlock extends FarmBlock
 	}
 
 	@Override
-	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
 		entity.causeFallDamage(fallDistance, 1.0F, entity.damageSources().fall());
 		// Rich Soil is immune to trampling
 	}
